rename match_stream -> match
